// src/api-client/models/payment-history-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface PaymentHistoryResponse
 */
export interface PaymentHistoryResponse {
    /**
     * 
     * @type {number}
     * @memberof PaymentHistoryResponse
     */
    'id'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PaymentHistoryResponse
     */
    'isPrimary'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof PaymentHistoryResponse
     */
    'paymentMethod'?: string;
    /**
     * 
     * @type {number}
     * @memberof PaymentHistoryResponse
     */
    'totalAmount'?: number;
    /**
     * 
     * @type {string}
     * @memberof PaymentHistoryResponse
     */
    'paymentStatus'?: PaymentHistoryResponsePaymentStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof PaymentHistoryResponse
     */
    'paymentTime'?: string;
}

export const PaymentHistoryResponsePaymentStatusEnum = {
    Paid: 'PAID',
    Pending: 'PENDING',
    Refunded: 'REFUNDED',
    Canceled: 'CANCELED',
    Failed: 'FAILED'
} as const;

export type PaymentHistoryResponsePaymentStatusEnum = typeof PaymentHistoryResponsePaymentStatusEnum[keyof typeof PaymentHistoryResponsePaymentStatusEnum];